+++ s/transports/messenger/conduits/postable.ts
@@ -1,18 +1,18 @@
 
 import {disposer} from "@e280/stz"
 import {Conduit} from "./conduit.js"
-import {PostableChannel} from "../types.js"
+import {Messagable} from "../types.js"
 import {onMessage} from "../parts/helpers.js"
 
 export class PostableConduit extends Conduit {
 	dispose = disposer()
 
-	constructor(channel: PostableChannel) {
+	constructor(messagable: Messagable) {
 		super()
 		this.dispose.schedule(
-			this.sendRequest.sub((m, transfer) => channel.postMessage(m, transfer)),
-			this.sendResponse.sub((m, transfer) => channel.postMessage(m, transfer)),
-			onMessage(channel, e => this.recv(e.data, e)),
+			this.sendRequest.sub((m, transfer) => messagable.postMessage(m, transfer)),
+			this.sendResponse.sub((m, transfer) => messagable.postMessage(m, transfer)),
+			onMessage(messagable, e => this.recv(e.data, e)),
 		)
 	}
 }