+++ s/transports/messenger/conduits/window.ts
@@ -1,7 +1,7 @@
 
 import {disposer} from "@e280/stz"
 import {Conduit} from "./conduit.js"
-import {ChannelMessage} from "../types.js"
+import {MessageLike} from "../types.js"
 import {onMessage} from "../parts/helpers.js"
 
 export class WindowConduit extends Conduit {
@@ -12,7 +12,7 @@ export class WindowConduit extends Conduit {
 			localWindow: Window
 			targetWindow: WindowProxy
 			targetOrigin: string
-			allow: (e: ChannelMessage) => boolean
+			allow: (e: MessageLike) => boolean
 		}) {
 
 		super()