+++ s/transports/messenger/types.ts
@@ -13,17 +13,28 @@ export type MessengerOptions<LocalFns extends Fns, RemoteFns extends Fns> = {
 	conduit: Conduit
 	tap?: Tap
 	timeout?: number
+
+	// TODO maybe rpc should be required?
 	rpc?: MessengerRpc<LocalFns, RemoteFns>
 }
 
-export type ChannelMessage<D = any> = {data: D, origin: string}
+export type MessageLike<D = any> = {data: D, origin: string, source: MessageEvent["source"]}
 
-export type Channel = {
-	addEventListener(e: "message", listener: (event: ChannelMessage) => void): void
-	removeEventListener(e: "message", listener: (event: ChannelMessage) => void): void
+export type MessageReceiver = {
+	addEventListener(e: "message", listener: (event: MessageLike) => void): void
+	removeEventListener(e: "message", listener: (event: MessageLike) => void): void
 }
 
-export type PostableChannel = {
+export type Messagable = {
 	postMessage(m: any, transfer?: Transferable[]): void
-} & Channel
+} & MessageReceiver
+
+/** @deprecated renamed to 'MessageLike' */
+export type ChannelMessage<D = any> = MessageLike<D>
+
+/** @deprecated renamed to 'MessageReceiver' */
+export type Channel = MessageReceiver
+
+/** @deprecated renamed to 'Messagable' */
+export type PostableChannel = Messagable